+++ applications/settings/cfw_app/cfw_app.c
@@ -62,7 +62,7 @@ bool cfw_app_apply(CfwApp* app) {
             while(flipper_format_delete_key(file, "Frequency"))
                 ;
             FrequencyList_it(it, app->subghz_static_freqs);
-            for(uint i = 0; i < FrequencyList_size(app->subghz_static_freqs); i++) {
+            for(size_t  i = 0; i < FrequencyList_size(app->subghz_static_freqs); i++) {
                 flipper_format_write_uint32(
                     file, "Frequency", FrequencyList_get(app->subghz_static_freqs, i), 1);
             }
@@ -70,7 +70,7 @@ bool cfw_app_apply(CfwApp* app) {
             if(!flipper_format_rewind(file)) break;
             while(flipper_format_delete_key(file, "Hopper_frequency"))
                 ;
-            for(uint i = 0; i < FrequencyList_size(app->subghz_hopper_freqs); i++) {
+            for(size_t  i = 0; i < FrequencyList_size(app->subghz_hopper_freqs); i++) {
                 flipper_format_write_uint32(
                     file, "Hopper_frequency", FrequencyList_get(app->subghz_hopper_freqs, i), 1);
             }
